# pkg/compute/openapi.yaml
openapi: 3.0.3
info:
  title: Compute API
  version: v1
  description: Internal REST API for sandbox, snapshot, template, and operation management.
servers:
  - url: /
...
paths:
  /v1/me/ssh-keys:
    get:
      tags: [ssh-keys]
      operationId: listSSHPublicKeys
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: SSH public keys registered by the authenticated subject.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SSHPublicKeyList'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [ssh-keys]
      operationId: addSSHPublicKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddSSHPublicKeyRequest'
      responses:
        '201':
          description: SSH public key registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SSHPublicKey'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/me/ssh-keys/{id}:
    parameters:
      - $ref: '#/components/parameters/SSHPublicKeyID'
    delete:
      tags: [ssh-keys]
      operationId: deleteSSHPublicKey
      responses:
        '204':
          description: SSH public key removed.
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/sandboxes:
    get:
      tags: [sandboxes]
      operationId: listSandboxes
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Sandboxes in stable ID order.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SandboxList'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [sandboxes]
      operationId: createSandbox
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSandboxRequest'
      responses:
        '202':
          description: Sandbox creation accepted. Poll the returned operation ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/sandboxes/{id}:
    parameters:
      - $ref: '#/components/parameters/SandboxID'
    get:
      tags: [sandboxes]
      operationId: getSandbox
      responses:
        '200':
          description: Sandbox.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sandbox'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [sandboxes]
      operationId: deleteSandbox
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: Sandbox deletion accepted. Poll the returned operation ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/sandboxes/{id}/stop:
    parameters:
      - $ref: '#/components/parameters/SandboxID'
    post:
      tags: [sandboxes]
      operationId: stopSandbox
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: Stop accepted. Poll the returned operation ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedResponse'
        default:
          $ref: '#/components/responses/Error'
  /v1/sandboxes/{id}/start:
    parameters:
      - $ref: '#/components/parameters/SandboxID'
    post:
      tags: [sandboxes]
      operationId: startSandbox
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: Start accepted. Poll the returned operation ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedResponse'
        default:
          $ref: '#/components/responses/Error'
  /v1/sandboxes/{id}/snapshot:
    parameters:
      - $ref: '#/components/parameters/SandboxID'
    post:
      tags: [sandboxes]
      operationId: snapshotSandbox
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SnapshotSandboxRequest'
      responses:
        '202':
          description: Snapshot accepted. Poll the returned operation ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedResponse'
        default:
          $ref: '#/components/responses/Error'
  /v1/sandboxes/{id}/lease:
    parameters:
      - $ref: '#/components/parameters/SandboxID'
    post:
      tags: [sandboxes]
      operationId: leaseSandbox
      summary: Extend a sandbox's TTL. A heartbeat is a periodic call to this endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeaseSandboxRequest'
      responses:
        '200':
          description: Lease extended. Returns the new expiry, clamped to max TTL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaseResponse'
        default:
          $ref: '#/components/responses/Error'
...
  /v1/snapshots:
    get:
      tags: [snapshots]
      operationId: listSnapshots
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Snapshots in stable ID order.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotList'
        default:
          $ref: '#/components/responses/Error'
  /v1/snapshots/{id}:
    parameters:
      - $ref: '#/components/parameters/SnapshotID'
    get:
      tags: [snapshots]
      operationId: getSnapshot
      responses:
        '200':
          description: Snapshot.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Snapshot'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [snapshots]
      operationId: deleteSnapshot
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: Snapshot deletion accepted. Poll the returned operation ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedResponse'
        default:
          $ref: '#/components/responses/Error'
  /v1/snapshots/{id}/restore:
    parameters:
      - $ref: '#/components/parameters/SnapshotID'
    post:
      tags: [snapshots]
      operationId: restoreSnapshot
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestoreSnapshotRequest'
      responses:
        '202':
          description: Restore accepted. Poll the returned operation ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedResponse'
        default:
          $ref: '#/components/responses/Error'
  /v1/templates:
    get:
      tags: [templates]
      operationId: listTemplates
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Templates in stable name/version order.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateList'
        default:
          $ref: '#/components/responses/Error'
  /v1/templates/{name}/{version}:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [templates]
      operationId: getTemplate
      responses:
        '200':
          description: Template.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/operations:
    get:
      tags: [operations]
      operationId: listOperations
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Operations in stable created_at/id order.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationList'
        default:
          $ref: '#/components/responses/Error'
  /v1/operations/{id}:
    parameters:
      - $ref: '#/components/parameters/OperationID'
    get:
      tags: [operations]
      operationId: getOperation
      responses:
        '200':
          description: Operation. Terminal states are succeeded and failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /v1/sandboxes/{id}/logs:
    parameters:
      - $ref: '#/components/parameters/SandboxID'
    get:
      tags: [sandboxes]
      operationId: getSandboxLogs
      responses:
        '200':
          description: Guest log lines (empty in deferred mode).
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /v1/sandboxes/{id}/lineage:
    parameters:
      - $ref: '#/components/parameters/SandboxID'
    get:
      tags: [sandboxes]
      operationId: getSandboxLineage
      responses:
        '200':
          description: Derived provenance graph for the sandbox.
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /v1/sandboxes/{id}/operations:
    parameters:
      - $ref: '#/components/parameters/SandboxID'
    get:
      tags: [sandboxes]
      operationId: getSandboxOperations
      responses:
        '200':
          description: Operations targeting the sandbox.
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /v1/sandboxes/{id}/snapshots:
    parameters:
      - $ref: '#/components/parameters/SandboxID'
    get:
      tags: [sandboxes]
      operationId: getSandboxSnapshots
      responses:
        '200':
          description: Snapshots taken from the sandbox.
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /v1/snapshots/{id}/restored-by:
    parameters:
      - $ref: '#/components/parameters/SnapshotID'
    get:
      tags: [snapshots]
      operationId: getSnapshotRestoredBy
      responses:
        '200':
          description: Sandboxes restored from the snapshot.
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /v1/nodes:
    get:
      tags: [system]
      operationId: listNodes
      responses:
        '200':
          description: Node registry entries.
        default:
          $ref: '#/components/responses/Error'
  /v1/nodes/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [system]
      operationId: getNode
      responses:
        '200':
          description: A node registry entry.
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /v1/users:
    get:
      tags: [system]
      operationId: listUsers
      responses:
        '200':
          description: Users.
        default:
          $ref: '#/components/responses/Error'
  /v1/users/{handle}:
    parameters:
      - name: handle
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [system]
      operationId: getUser
      responses:
        '200':
          description: A user.
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
...
  /v1/audit:
    get:
      tags: [system]
      operationId: listAudit
      responses:
        '200':
          description: Append-only audit log entries.
        default:
          $ref: '#/components/responses/Error'
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Required for mutating requests. Reusing a key with the same request returns the same accepted operation; reusing it with a different request returns 409.
      schema:
        type: string
        minLength: 1
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 100
    Cursor:
      name: cursor
      in: query
      required: false
      description: Integer offset cursor returned as next_cursor.
      schema:
        type: string
    Offset:
      name: offset
      in: query
      required: false
      description: Integer offset alternative to cursor.
      schema:
        type: integer
        minimum: 0
    SandboxID:
      name: id
      in: path
      required: true
      schema:
        type: string
...
    SSHPublicKeyID:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    Error:
      description: API error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    BadRequest:
      description: Bad request, including malformed JSON, missing idempotency key, or invalid pagination parameters.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: Conflict, including idempotency-key reuse with a different request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    UnprocessableEntity:
      description: Syntactically valid request that failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    InternalError:
      description: Internal server error. Message is masked and does not include internal details.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    CreateSandboxRequest:
      type: object
      required: [template]
      additionalProperties: false
      properties:
        template:
          type: string
        ttl:
          type: string
          description: Optional positive Go-style duration string.
        on_delete:
          type: string
          enum: [delete, hot, cold, archive]
          description: Disposition to apply when the sandbox is deleted.
    SnapshotSandboxRequest:
      type: object
      additionalProperties: false
      properties:
        note:
          type: string
...
    AddSSHPublicKeyRequest:
      type: object
      required: [public_key]
      additionalProperties: false
      properties:
        name:
          type: string
          description: Optional display name. Defaults to the key comment or fingerprint.
        public_key:
          type: string
          description: One OpenSSH public key line.
    SSHPublicKey:
      type: object
      required: [id, name, public_key, fingerprint, created_at, updated_at]
      properties:
        id:
          type: string
        name:
          type: string
        public_key:
          type: string
        fingerprint:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AcceptedResponse:
      type: object
      required: [op_id]
      properties:
        id:
          type: string
          description: Resource ID for create-style operations.
        sandbox_id:
          type: string
        snapshot_id:
          type: string
        op_id:
          type: string
...
    SandboxList:
      allOf:
        - $ref: '#/components/schemas/ListEnvelope'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Sandbox'
...
    SSHPublicKeyList:
      allOf:
        - $ref: '#/components/schemas/ListEnvelope'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/SSHPublicKey'
...